feat: implement 5-minute payment timeout, database notification channel, and class-based broadcasting

+++ database/migrations/0001_01_01_000044_create_carts_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->id();
             $table->unsignedBigInteger('user_id');
             $table->unsignedBigInteger('vendor_inventory_id');
-            $table->unsignedBigInteger('quantity');
+            $table->decimal('quantity');
             $table->unsignedBigInteger('cart_status_id');
             $table->timestamps();
             $table->softDeletes();
